Move all highlighting and playing logic in the player
This means all the compuation logic in the song state is no longer
necessary. Only the active bar is now highlighted instead of the active
strum. It's possible to change the bpm of the song playback and the
highlighting is still synchronized well. This also solves the bug that
the highliht and the player can get out of sync if playback is slow.