Cleanup and fixing tests
It's now allowed to use [instrument] in instrument parsing to specify
the instrument type. Before this caused issues if there was more than
one way to parse the instrument (bass is an instrument and a sound).

Patterns require the leading and closing bar `|`. This solves issues
when a pattern name ends with a number.